Modifies the scoring algorithm
The new scoring algorithm searches the entire board for all positions
where a win could occur. If the computer wins, the maximum allowble score
is used. If it loses, the negative of the maximum allowable score is used.
If neither player wins, the score for each possible winning position is
the number of pieces the AIPlayer has in the individual spots of the
overall required position to win. The score of the state is the sum of the
scores of all possible winning positions.